+++ src/main/java/com/hyerijang/dailypay/auth/controller/AuthenticationController.java
@@ -1,5 +1,9 @@
-package com.hyerijang.dailypay.auth;
+package com.hyerijang.dailypay.auth.controller;
 
+import com.hyerijang.dailypay.auth.dto.AuthenticationRequest;
+import com.hyerijang.dailypay.auth.dto.AuthenticationResponse;
+import com.hyerijang.dailypay.auth.dto.RegisterRequest;
+import com.hyerijang.dailypay.auth.service.AuthenticationService;
 import com.hyerijang.dailypay.common.aop.ExeTimer;
 import com.hyerijang.dailypay.common.exception.ApiException;
 import com.hyerijang.dailypay.common.exception.response.ExceptionEnum;
