+++ src/main/java/com/hyerijang/dailypay/auth/dto/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package com.hyerijang.dailypay.auth;
+package com.hyerijang.dailypay.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
